Added shards Moved index.js to bot.js Created new index.js with a shardmanager Added shard command to check shards and how many guilds they have
