+++ index.js
@@ -1,47 +1,8 @@
-const fs = require('fs');
-const { Client, Collection, Intents } = require('discord.js');
+const { ShardingManager } = require('discord.js');
 const { token } = require('./config.json');
-const { errembed } = require('./functions');
 
-// Create a client with the required intents
-const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MEMBERS] });
+const manager = new ShardingManager('./bot.js', { token: token });
 
-client.commands = new Collection();
-const commandFiles = fs.readdirSync('./commands').filter(file => file.endsWith('.js'));
+manager.on('shardCreate', shard => console.log(`Launched shard ${shard.id}`));
 
-for (const file of commandFiles) {
-	const command = require(`./commands/${file}`);
-	// Set a new item in the Collection
-	// With the key as the command name and the value as the exported module
-	client.commands.set(command.data.name, command);
-}
-
-const eventFiles = fs.readdirSync('./events').filter(file => file.endsWith('.js'));
-
-for (const file of eventFiles) {
-	const event = require(`./events/${file}`);
-	if (event.once) {
-		client.once(event.name, (...args) => event.execute(...args));
-	} else {
-		client.on(event.name, (...args) => event.execute(...args));
-	}
-}
-
-client.on('interactionCreate', async interaction => {
-	if (!interaction.isCommand()) return;
-
-	const command = client.commands.get(interaction.commandName);
-
-	if (!command) return;
-
-	try {
-		await command.execute(interaction);
-	} catch (error) {
-		console.error(error);
-		// Create a new error embed using the function from functions.js
-		errembed({ interaction: interaction, author: 'There was an error while executing this command!' });
-	}
-});
-
-// Login to the client using the token from config.json
-client.login(token);
+manager.spawn();
